Initial Commit.
I decided to upload at the end of each project my web development code from a course I am currently taking so that I can be more attractive in the eyes of a future employer.

The first project that this contains transitions between active images which take up most of the width which is decided by what you click on at which point a CSS transition effect occurs to expand the new active image and collapse the previously active panel.